feat: adjust weighting
Manually set the weights of different motions to hopefully get more sensible
results. For example, only hjkl have a weight of 1. This means that where
there is a choice between using e.g. ^ or h to move a single character, h will
take precedence since it makes more sense to the user.

Weights are mainly based on the number of characters typed and the complexity
of the motion (simpler movements are easier to understand).