feat: calculate rweight based on effect on count
Instead of using a fixed weight for each subsequent repetition of a
motion, calculate the cost of adding another repetition using the change
in length of the counter. For example:

MOTION    COST OF ADDING REPETITION
j -> 2j   1 (2j is 1 char longer than j)
2j -> 3j  0 (3j is no longer than 2j)
9j -> 10j 1 (10j is 1 char longer than 9j)

The initial usage still has a configurable cost.