feat: refine the path during backtracking
line

para
graph

parag
raph

In the example above, the start point is on `line` and the target is a blank
line after the second paragraph. For simplicity pretend that the only
available motions are j and }.

Preiously, the pathfinder would have returned j2}. This is because j comes
first in the settings and when } is tested, it goes to the same location with
the same cost and was ignored.

Now, both j and } will be stored in a list called `incoming_motions` attached
to that node. When we reach the target, we go back and choose the best
motion from each of those lists, meaning that 3} will be the output.

Fixes #10