perf: run pathfinding in a separate Vim instance
Do pathfinding in a subprocess which runs a "server" within a stripped-down,
plugin-free Vim. This increases the speed of pathfinding significantly.
It also leaves the user's cursor alone, and doesn't block - so they may
continue working. This is necessary to implement #12.

The client is responsible for launching and connecting to the server.
The server will shut down it's Vim process when the client disconnects.

The server starts using a custom vimrc (serverrc.vim), which disables
unnecessary features like syntax highlighting and other plugins.
pathfinder.vim checks for the existance of a g:pf_server_communiation_file
variable (which the client sets using --cmd), if that exists it starts in
server mode, otherwise it starts in client mode.

Communication between the client (user's Vim) and server (subprocess Vim)
happens with a temporary file and multiprocessing.connection classes.
The current buffer contents, g:pf_motions setting, scrolloff setting, start
and target positions and the window size are transferred to the server
before each pathfinding request. The window size is restored using
&lines and &columns, meaning the surrounding windows don't need to be
transferred.