feat: asynchronous and automatic pathfinding (#13)
* refactor: convert path into a class

* perf: run pathfinding in a separate Vim instance

Do pathfinding in a subprocess which runs a "server" within a stripped-down,
plugin-free Vim. This increases the speed of pathfinding significantly.
It also leaves the user's cursor alone, and doesn't block - so they may
continue working. This is necessary to implement #12.

The client is responsible for launching and connecting to the server.
The server will shut down it's Vim process when the client disconnects.

The server starts using a custom vimrc (serverrc.vim), which disables
unnecessary features like syntax highlighting and other plugins.
pathfinder.vim checks for the existance of a g:pf_server_communiation_file
variable (which the client sets using --cmd), if that exists it starts in
server mode, otherwise it starts in client mode.

Communication between the client (user's Vim) and server (subprocess Vim)
happens with a temporary file and multiprocessing.connection classes.
The current buffer contents, g:pf_motions setting, scrolloff setting, start
and target positions and the window size are transferred to the server
before each pathfinding request. The window size is restored using
&lines and &columns, meaning the surrounding windows don't need to be
transferred.

* feat: run pathfinder automatically

BREAKING CHANGE: The `:PathfinderBegin` and `:PathfinderRun` commands have been removed.

Pathfinding runs when...
- The cursor has not moved for a configurable (g:pf_autorun_delay) time
  (default 2 seconds)
  > End position is the current cursor position
- An edit (x,rx,p etc.) was made in normal mode
  > End position is the position just before the edit
- Switching modes from normal, visual or visual-line
  > End position is the position just before switching modes

The start position is wherever the cursor was after...
- The pathfinder last ran
- Switching window or tab
- Opening a file or an empty buffer
- Loading a session

* feat: display output in popup if possible

If Vim has +popupwin (available in 8.2), use that to display the full
output in a box at the bottom of the screen.

Otherwise, display a single-line compact form using echo. (Having only
one line prevents a "press enter" prompt from appearing.)

* feat: scale popup time based on length of output

Increase the time the popup is shown for when it has a lot of lines, so
that it doesn't disappear too fast for the user to read it.

Suggestion: make this configurable.

* docs: update README