perf: limit search area to around the start and target
Disallow the exploration of nodes outside the search area. The search
area is defined as the lines between the start and target nodes, plus a
number of lines above and below. The number of extra lines is based on
the length of the movement - moving further means more surrounding lines
will be considered.

By preventing Dijkstra's algorithm from exploring outside of this area,
performance is significantly improved, although sometimes a better path
is sacrificed. However, that better path would most often not be obvious
to a human without first scrolling the window up and down to see the
surrounding lines. The better speed is most noticeable when making
complicated horizontal movements on a single line.

Fixes #7