feat: make expore_lines configurable
g:pf_explore_scale sets the value the length of the search area is
multiplied by to get the number of lines to explore either side.
g:pf_max_explore sets a maximum value this will be capped at.

These settings default to 0.5 and 10. This means that if the search area
(in between the start and target positions) was 10 lines long, 5 more
would be explored above and below. But if the search area was 30 lines
long, only 10 would be explored either side despite 30*0.5 being 15,
because of the limit.