perf: limit search area to around the start and target (#23)
Disallow the exploration of nodes outside a certain area around the
start and target positions. By preventing the algorithm from exploring
outside of this area, performance is significantly improved, although
sometimes a better path is sacrificed. However, that better path would
most often be non-obvious to a human.

g:pf_explore_scale sets the value the length of the search area (lines
between and including the start and target positions) is multiplied by to
get the number of lines to explore either side.
g:pf_max_explore sets a maximum value this will be capped at.

Fixes #7