perf: use heapdict for a faster priority queue
This module provides faster updating of node priorities than heapq,
which improves performance.

Although it can be installed through pip, this has been added as a Git
submodule because a lot of plugin managers will know to download it
without the user needing to set a post-install command or run pip manually.