feat: display suggestion close to cursor
Move the suggestion right next to the cursor, so it can be read without
having to look to the bottom of the screen (for people with +popupwin,
anyway).

BREAKING CHANGE: `PathfinderPopup` highlight group is no longer used.
The popup now inherits the appearance of the cursor.

BREAKING CHANGE: From now on, suggestions will always appear in compact form.
A command, `:PathfinderExplain`, has been added to view descriptions.