perf: rewrite pathfinding algorithm
The previous algorithm was very overcomplicated. It is possible to
simply create a separate vertex in the graph for each previously-used
motion, removing the need for the "refining" stage. This also allows
consideration of whole congruent paths rather than only a single motion
with the same effect.

Also removed the Node and StartNode classes, replacing them with
named tuples.